drf-intro/simple_crud/measurements/views.py
from rest_framework.viewsets import ModelViewSet

from .models import Measurement, Project
from .serializers import MeasurmentSerializer, ProjectSerializer

class ProjectViewSet(ModelViewSet):
    """ViewSet для проекта."""
    # TODO: добавьте конфигурацию для объекта
    queryset = Project.objects.all()
    serializer_class = ProjectSerializer


class MeasurementViewSet(ModelViewSet):
    """ViewSet для измерения."""
    # TODO: добавьте конфигурацию для измерения
    queryset = Measurement.objects.all()
    serializer_class = MeasurmentSerializer
